A-sko/requirements.txt
Flask
Flask-Login
Werkzeug
snowflake-connector-python